Tidiness of content script re-injection
Filter tabs' URLs so we don't try to inject the content script into
inappropriate tabs. This is thwarted somewhat by Chrome's special-casing
of the Chrome Web Store (which it calls the Extensions Gallery), so an
error is raised in this case, and there doesn't seem to be a different
way to detect if the user is browsing the Chrome Web Store.