Initial profiling support
Have tried many different numbers of repetitions, wait times and steps
paramater for the mouse move, but only ever showing one or two (very
rarely four) calls to shouldRefreshLandmarks() and not really sure why.

Actually, on further testing, it seems that the number of calls to
shouldRefreshLandmarks() varies across runs of the trace. Nowhere near
as many calls as actually happen are picked up (assuming this is down to
timing resolution), but the number picked up does vary across runs.

Still, the main purpose of this is to streamline the functions that take
the longest, i.e. the landmarks-finding code. Would like to grok this
whole tracing thing, though...