Specify unused relevant CSS; Remove data attribute
This resets various CSS properties that are not explicitly used by the
border and label elements, in case the page does set them. It also
removes a redundant data attribute that was used early on in #158 but a
better way was found to ignore the borders so it's no longer needed.