Make the inspect buttons work
This probably is not 100% robust, but has worked everywhere I have tried
it so far. It occurred to me that it may be easier than I first thought
to generate a reliable selector because it can be entirely specific. One
way might have been to use nth-child() a lot, but that would lead to
long strings. This way is simple and works well, though one area it may
fall over is that it does not use nth-child at all. Need some test
cases, but it feels promising.

Minor code reorganisation, such as for button-making, based on these
changes.

The inspect button now also gives context in its ARIA label, and has a
title (thus tooltip and accessible description) that shows the computed
selector.

Still to do: update comments and user docs.

Note: this was committed without tests because there is now an extra
field added to the element info objects returned by the LandmarksFinder.
Need to think about whether to add the selectors to the existing tests
(seems most proper) or whether to only test, or request, the selectors
when needed (may improve speed, but premature optimisation?)