Warn user about DevTools disconnect; messages file
On Chrome and Opera, the DevTools page is not reloaded when the rest of
the extension is. I tried a few ways to mitigate this (trying to
reconnect to the background script; trying to reload the DevTools page)
but the code was both not neat and actually did not work, so have
resorted to simply warning the user for now, which is probably a good
idea as it will not happen often.

This also fixes some missing blank lines in the messages file.