Fix file missing from zip
The changes in 3417085160ac9115362af4054eb03d6f6aa008a9 borked things in
an odd way: devtoolsPanel.js was missing from the created zip files.

Initially I tried fixing things by restoring the natural async behaviour
of the zip creation and linting only after the zip has made. Output from
the build script is not as neat.

This does make one change to how archiver is used: it passes false
instead of an empty string as the path under which the build directory
tree is to be added to the zip file. This is in line with the latest
archiver README.

That alone didn't actually solve things. I then changed the rollup code
to use await rather than promises for writing the bundle (as per the
current code examples on the rollup site). Now it's working again.

Decided not to try going back to the previous archiver code, as not
really sure if what I was doing regarding the handling of promises was
right (didn't seem great, mixed in with async/await).