Multiple run support for profiling script
When running it, tell the profiling script how many runs to, erm, run.
This has shown, though, that it seems using the browser for performance
testing like this is not sufficiently reproducible. Was thinking of just
starting the trace later, and also tried letting the page take longer to
settle, in case it perturbed the cadence of adding the landmarks later,
but neither seemed to help.