Streamline message dismissal
A message appears in the sidebar when it is open, but is not the default
interface. The message can be dismissed. There is a button in the
options page to reset the dismissal states. This patch:

* Makes it so that the button to reset dismissed messages is enabled or
  disabled based on if any messages have been dismissed (instead of always
  being enabled and using an alert to notify the user that messages have
  been reset - prompted because Firefox now makes the alert transparent
  for some reason).
* Automatically re-instates the sidebar message as soon as the dismissal
  states are reset.
* Only automatically re-instates the sidebar message if the sidebar is
  still not the primary interface the user has chosen.

The only downside is that the button, being disabled, will not be in the
focus order, so not easily discoverable by people who can't see the
page.

Partly addresses #221.