feat: Enhanced help documentation, bundled with the extension (#237)
Add a new, HTML help page that explains how to use the extension, including reflecting keyboard shortcuts and providing links to go directly to the settings, for convenience.

This was created by moving some content from the README and the techniques from the original “splash” content script. It also directly imports the main Landmarks content script so that the extension can itself be used to navigate this page. Refer to #165 for the details.

Partly addresses #165 (provides the foundation for managing the keyboard shortcuts when another one is added).