Use "refresh" instead of "update"...
...to make it clear that after a change to the document, nothing would
automatically be added, but the current state reflected as far as
possible (i.e. removing border on focused element if it has gone;
updating all borders).

Currently it does not remove borders on elements that no longer exist
when it is trying to resize the borders. This will lead to errors, but
there is already a FIXME about it.