feat(toggle): Show all landmarks keyboard shortcut (#245)
Naughty! This really should’ve been split up into smaller PRs, though I think I am starting to get into this new commit style :-). Anyway, this does a couple of things:

* Adds help page support for indicating to the user when a keyboard shortcut is not set (in support of #165).
* Implements a keyboard shortcut that toggles the display of all landmarks (in support of #120).

In support of this, it required splitting the ElementFocuser into an ElementFocuser and a BorderDrawer. A lot of API naming cleanup and code tidying was done too.

A further PR will add UI to the pop-up that mouse users can use to access this feature.