fix(UI): Add help and preferences links to GUIs (#253)
This makes Help and Preferences more discoverable. It involved renaming some messages for clarity, factoring out translation and also improving the styles somewhat, including having slight differences on the sidebar styles (to make it fill height). The links don't show up in the DevTools panel.

When the stand-alone options page is done, there should be a big refactoring and harmonisation of styles across options, GUIs and help.

Fixes #250.