fix: Move to simpler message passing (#268)
This removes as much of the long-lived port connections as possible - i.e. those between the main GUIs (but DevTools requires them, so there they stay). Amount of code is reduced and it does feel much simpler (fewer cross-browser workarounds are required, too).

It is possible to listen for navigation-related events in most GUIs, so the background script doesn't need to act on their behalf, but not for DevTools, so the background script would still need to do that. Therefore it's neater that that stuff still happens in the background script.

Fixes #265