Dark mode; Remove pop-up heading; General styling
I did try using the browsers' themes for the popup and sidebar, but the
backgrounds didn't reflect dark mode, and there's no API in Chrome to
query for the current theme colours.

Generally the styles have been tweaked, streamlined and harmonised with
the styles used for the help and new options pages.