Windows-related and browser style fixes
Had to revert to using browser styles in pop-ups and sidebars in all browsers. This is because I'd misunderstood that Chrome makes it optional in pop-ups (it doesn't, and there's no such option) and this caused lots of issues with font sizing across browsers. Now it is more consistent and better for users.

Also think I finally have a good fix for the pop-up sizing: don't prescribe a magic number - say how we want the content to look (in this case: don't break the 'show all' label) :-).

Not sure why the package-lock is different on Windows :-S.