Be more mobile-first
Improve layout on small screens.

Note: on my iPhone 6S, as I use Zoomed view, the <h1> is still a little
bit too wide for the screen, but not doing anything about this at the
moment, as it looks fine in normal mode and capping the heading size to
the viewport width seems like a bad idea (and I can't think of other
options at the mo).

Fixes #343