fix: Check dependency tree for modifications when caching code (#359)
Whoops; premature optimisation :-). I was not considering whether the sources on which a 'main' source file _depends_ had changed, when checking the modification dates in the code-caching bit. Luckily found a nice package to help me do this, and got to practice being a bit functional.

Had been thinking of renaming some variables for clarity, and "bundle" sounds nicer than "flatten", so have made that change here too.