Streamline scanner-changing
Have tested for a while and not come across a situation in which the
scanner was null when asked for a scan, but also there is no guarantee
that webNavigation.onCompleted-resulting message will come after the
devtools-state one (though content scripts do seem to start up before
the background script notices, which seems nice).

Anyway this guards against that issue by defaulting to the standard
scanner (again-ish) and also removes checks for trying to change to the
same scanner, as they do not seem to be needed.