feat: Guess the "main" region if none exists (#430)
If no "main" region is found, then look for some common IDs/class names to see if we can guess it. This partly addresses #181.

This adds a test suite for the new feature, and support for it in the profiling script.

This does support situations where there are other landmarks, but it assumes that the "main" region is at the top level of the tree.