build: Switch to ES modules (#436)
Node 14, which is currently LTS, has good ES module support, and adopting it will allow me to use the `??` operator. Seems like maybe quite a bit of work for just that :-) but this has simplified/consolidated/harmonised things a bit too.

This doesn't change the end product of the build, which still uses rollup to create a flattened version of the code.