fix: Ignore guessed regions with no innerText (#445)
I noticed a site in the wild with several guessed regions, including
some spurious ones. This addresses some of the issue by ignoring any
region that's essentially empty from the user's perspective. It does not
address regions that are visually hidden, as getComputedStyle() doesn't
cover hidden parent elements. Have decided not to change that for now,
due to the expected performance hit and lack of other examples of false
positives.