docs: Notes for 2.10.1 (#453)
This removes keyboard shortcut management instructions for non-Firefox
browsers. This is because I noticed that Operas UI has changed and
realised that they all could, so it would be easy for these to go
out-of-date, and we have the buttons to open the settings on non-Firefox
browsers anyway (which will need continued testing).