feat: Add an option to control whether heuristics are used (#459)
This addresses makes the use of heuristics an option that is on by default. It also re-organises the preferences page a bit to accommodate this. The radio buttons and checkboxes are enlarged in line with the rest of the text on the preferences page.

The content script code was cleaned up a little.

Fixes #457