Check for successful mutation handling
This checks that a full scan yields the same results as the specific
mutation handling did (currently trivial, as the mutation-specific stuff
has not been written yet).

It also checks whether the mutated/full and reverted results are the
same. Sometimes they should be, sometimes they should not.

Checks only run on first iteration, as before.