fix: Don't un-highlight persistent border (#487)
Don't hide the border of the current element when highlighting or
un-highlighting via the pop-up when borders are persistent.

Note: if the user has the mouse hovered over a button when the pop-up is
closed by pressing the Escape key, that one will remain highlighted. The
blur event fires on the button if focused, but apparently the mouseleave
event doesn't. It also doesn't seem possible to detect the pop-up window
closing.